added debug messages when error occur
