+++ .gitignore
@@ -1,0 +1,27 @@
+/autom4te.cache/
+/m4/
+/.libs/
+/Makefile
+/Makefile.in
+/aclocal.m4
+/compile
+/config.guess
+/config.log
+/config.status
+/config.sub
+/configure
+/depcomp
+/install-sh
+/libnfcauth.la
+/libtool
+/ltmain.sh
+/missing
+/nfcauth.lo
+/nfcauth.o
+/nfcauth_backend_config.lo
+/nfcauth_backend_config.o
+/pam-nfc-add
+/pam-nfc-add.o
+/pam_nfc.la
+/pam_nfc.lo
+/pam_nfc.o
